+++ src/directives/mwlCalendar.js
@@ -4,7 +4,7 @@ var angular = require('angular');
 
 angular
   .module('mwl.calendar')
-  .controller('MwlCalendarCtrl', function($scope, $log, $timeout, $attrs, $locale, moment, calendarTitle) {
+  .controller('MwlCalendarCtrl', function($scope, $log, $timeout, $attrs, $locale, moment, calendarTitle, calendarHelper) {
 
     var vm = this;
 
@@ -12,10 +12,10 @@ angular
 
     vm.changeView = function(view, newDay) {
       vm.view = view;
-      vm.currentDay = newDay;
+      vm.viewDate = newDay;
     };
 
-    vm.drillDown = function(date) {
+    vm.dateClicked = function(date) {
 
       var rawDate = moment(date).toDate();
 
@@ -25,13 +25,13 @@ angular
         week: 'day'
       };
 
-      if (vm.onDrillDownClick({calendarDate: rawDate, calendarNextView: nextView[vm.view]}) !== false) {
+      if (vm.onViewChangeClick({calendarDate: rawDate, calendarNextView: nextView[vm.view]}) !== false) {
         vm.changeView(nextView[vm.view], rawDate);
       }
 
     };
 
-    var previousDate = moment(vm.currentDay);
+    var previousDate = moment(vm.viewDate);
     var previousView = vm.view;
 
     function eventIsValid(event) {
@@ -58,7 +58,7 @@ angular
     function refreshCalendar() {
 
       if (calendarTitle[vm.view] && angular.isDefined($attrs.viewTitle)) {
-        vm.viewTitle = calendarTitle[vm.view](vm.currentDay);
+        vm.viewTitle = calendarTitle[vm.view](vm.viewDate);
       }
 
       vm.events = vm.events.filter(eventIsValid).map(function(event, index) {
@@ -67,7 +67,7 @@ angular
       });
 
       //if on-timespan-click="calendarDay = calendarDate" is set then don't update the view as nothing needs to change
-      var currentDate = moment(vm.currentDay);
+      var currentDate = moment(vm.viewDate);
       var shouldUpdate = true;
       if (
         previousDate.clone().startOf(vm.view).isSame(currentDate.clone().startOf(vm.view)) &&
@@ -87,52 +87,57 @@ angular
       }
     }
 
-    var eventsWatched = false;
+    calendarHelper.loadTemplates().then(function() {
+      vm.templatesLoaded = true;
 
-    //Refresh the calendar when any of these variables change.
-    $scope.$watchGroup([
-      'vm.currentDay',
-      'vm.view',
-      'vm.cellIsOpen',
-      function() {
-        return moment.locale() + $locale.id; //Auto update the calendar when the locale changes
-      }
-    ], function() {
-      if (!eventsWatched) {
-        eventsWatched = true;
-        //need to deep watch events hence why it isn't included in the watch group
-        $scope.$watch('vm.events', refreshCalendar, true); //this will call refreshCalendar when the watcher starts (i.e. now)
-      } else {
-        refreshCalendar();
-      }
+      var eventsWatched = false;
+
+      //Refresh the calendar when any of these variables change.
+      $scope.$watchGroup([
+        'vm.viewDate',
+        'vm.view',
+        'vm.cellIsOpen',
+        function() {
+          return moment.locale() + $locale.id; //Auto update the calendar when the locale changes
+        }
+      ], function() {
+        if (!eventsWatched) {
+          eventsWatched = true;
+          //need to deep watch events hence why it isn't included in the watch group
+          $scope.$watch('vm.events', refreshCalendar, true); //this will call refreshCalendar when the watcher starts (i.e. now)
+        } else {
+          refreshCalendar();
+        }
+      });
+
+    }).catch(function(err) {
+      $log.error('Could not load all calendar templates', err);
     });
 
   })
-  .directive('mwlCalendar', function(calendarUseTemplates) {
+  .directive('mwlCalendar', function(calendarConfig) {
 
     return {
-      template: calendarUseTemplates ? require('./../templates/calendar.html') : '',
-      restrict: 'EA',
+      templateUrl: calendarConfig.templates.calendar,
+      restrict: 'E',
       scope: {
         events: '=',
         view: '=',
         viewTitle: '=?',
-        currentDay: '=',
-        editEventHtml: '=',
-        deleteEventHtml: '=',
-        cellIsOpen: '=',
+        viewDate: '=',
+        editEventHtml: '=?',
+        deleteEventHtml: '=?',
+        cellIsOpen: '=?',
         onEventClick: '&',
         onEventTimesChanged: '&',
         onEditEventClick: '&',
         onDeleteEventClick: '&',
         onTimespanClick: '&',
-        onDrillDownClick: '&',
+        onViewChangeClick: '&',
         cellModifier: '&',
         dayViewStart: '@',
         dayViewEnd: '@',
-        dayViewSplit: '@',
-        monthCellTemplateUrl: '@',
-        monthCellEventsTemplateUrl: '@'
+        dayViewSplit: '@'
       },
       controller: 'MwlCalendarCtrl as vm',
       bindToController: true