+++ src/services/calendarHelper.js
@@ -4,7 +4,7 @@ var angular = require('angular');
 
 angular
   .module('mwl.calendar')
-  .factory('calendarHelper', function(dateFilter, moment, calendarConfig) {
+  .factory('calendarHelper', function($q, $templateRequest, dateFilter, moment, calendarConfig) {
 
     function formatDate(date, format) {
       if (calendarConfig.dateFormatter === 'angular') {
@@ -22,16 +22,15 @@ angular
       return moment(oldEnd).add(diffInSeconds);
     }
 
-    function eventIsInPeriod(event, periodStart, periodEnd) {
+    function getRecurringEventPeriod(eventPeriod, recursOn, containerPeriodStart) {
 
-      var eventStart = moment(event.startsAt);
-      var eventEnd = moment(event.endsAt || event.startsAt);
-      periodStart = moment(periodStart);
-      periodEnd = moment(periodEnd);
+      var eventStart = moment(eventPeriod.start);
+      var eventEnd = moment(eventPeriod.end);
+      var periodStart = moment(containerPeriodStart);
 
-      if (angular.isDefined(event.recursOn)) {
+      if (angular.isDefined(recursOn)) {
 
-        switch (event.recursOn) {
+        switch (recursOn) {
           case 'year':
             eventStart.set({
               year: periodStart.year()
@@ -46,12 +45,25 @@ angular
             break;
 
           default:
-            throw new Error('Invalid value (' + event.recursOn + ') given for recurs on. Can only be year or month.');
+            throw new Error('Invalid value (' + recursOn + ') given for recurs on. Can only be year or month.');
         }
 
-        eventEnd = adjustEndDateFromStartDiff(event.startsAt, eventStart, eventEnd);
+        eventEnd = adjustEndDateFromStartDiff(eventPeriod.start, eventStart, eventEnd);
 
       }
+
+      return {start: eventStart, end: eventEnd};
+
+    }
+
+    function eventIsInPeriod(event, periodStart, periodEnd) {
+
+      periodStart = moment(periodStart);
+      periodEnd = moment(periodEnd);
+
+      var eventPeriod = getRecurringEventPeriod({start: event.startsAt, end: event.endsAt || event.startsAt}, event.recursOn, periodStart);
+      var eventStart = eventPeriod.start;
+      var eventEnd = eventPeriod.end;
 
       return (eventStart.isAfter(periodStart) && eventStart.isBefore(periodEnd)) ||
         (eventEnd.isAfter(periodStart) && eventEnd.isBefore(periodEnd)) ||
@@ -88,11 +100,11 @@ angular
       return weekdays;
     }
 
-    function getYearView(events, currentDay, cellModifier) {
+    function getYearView(events, viewDate, cellModifier) {
 
       var view = [];
-      var eventsInPeriod = getEventsInPeriod(currentDay, 'year', events);
-      var month = moment(currentDay).startOf('year');
+      var eventsInPeriod = getEventsInPeriod(viewDate, 'year', events);
+      var month = moment(viewDate).startOf('year');
       var count = 0;
       while (count < 12) {
         var startPeriod = month.clone();
@@ -116,11 +128,11 @@ angular
 
     }
 
-    function getMonthView(events, currentDay, cellModifier) {
+    function getMonthView(events, viewDate, cellModifier) {
 
-      var startOfMonth = moment(currentDay).startOf('month');
+      var startOfMonth = moment(viewDate).startOf('month');
       var day = startOfMonth.clone().startOf('week');
-      var endOfMonthView = moment(currentDay).endOf('month').endOf('week');
+      var endOfMonthView = moment(viewDate).endOf('month').endOf('week');
       var eventsInPeriod;
       if (calendarConfig.displayAllMonthEvents) {
         eventsInPeriod = filterEventsInPeriod(events, day, endOfMonthView);
@@ -132,7 +144,7 @@ angular
 
       while (day.isBefore(endOfMonthView)) {
 
-        var inMonth = day.month() === moment(currentDay).month();
+        var inMonth = day.month() === moment(viewDate).month();
         var monthEvents = [];
         if (inMonth || calendarConfig.displayAllMonthEvents) {
           monthEvents = filterEventsInPeriod(eventsInPeriod, day, day.clone().endOf('day'));
@@ -161,10 +173,10 @@ angular
 
     }
 
-    function getWeekView(events, currentDay, filterOneDayEvents) {
+    function getWeekView(events, viewDate) {
 
-      var startOfWeek = moment(currentDay).startOf('week');
-      var endOfWeek = moment(currentDay).endOf('week');
+      var startOfWeek = moment(viewDate).startOf('week');
+      var endOfWeek = moment(viewDate).endOf('week');
       var dayCounter = startOfWeek.clone();
       var days = [];
       var today = moment().startOf('day');
@@ -181,20 +193,20 @@ angular
         dayCounter.add(1, 'day');
       }
 
-      if (filterOneDayEvents) {
-        events = events.filter(function(event) {
-          return !moment(event.startsAt).isSame(moment(event.endsAt), 'day');
-        });
-      }
-
       var eventsSorted = filterEventsInPeriod(events, startOfWeek, endOfWeek).map(function(event) {
 
-        var eventStart = moment(event.startsAt).startOf('day');
-        var eventEnd = moment(event.endsAt || event.startsAt).startOf('day');
         var weekViewStart = moment(startOfWeek).startOf('day');
         var weekViewEnd = moment(endOfWeek).startOf('day');
-        var offset, span;
 
+        var eventPeriod = getRecurringEventPeriod({
+          start: moment(event.startsAt).startOf('day'),
+          end: moment(event.endsAt || event.startsAt).startOf('day')
+        }, event.recursOn, weekViewStart);
+
+        var eventStart = eventPeriod.start;
+        var eventEnd = eventPeriod.end;
+
+        var offset, span;
         if (eventStart.isBefore(weekViewStart) || eventStart.isSame(weekViewStart)) {
           offset = 0;
         } else {
@@ -221,60 +233,23 @@ angular
 
     }
 
-    function getCrossingsCount(event, dayEvents) {
-      var eventStart = moment(event.startsAt);
-      var eventEnd = moment(event.endsAt);
+    function getDayView(events, viewDate, dayViewStart, dayViewEnd, dayViewSplit) {
 
-      return dayEvents.filter(function(ev) {
-
-        return event.$id !== ev.$id &&
-          (moment(ev.startsAt).isBetween(eventStart, eventEnd) ||
-          moment(ev.startsAt).isSame(eventStart) ||
-          moment(ev.endsAt).isBetween(eventStart, eventEnd) ||
-          moment(ev.endsAt).isSame(eventEnd) ||
-          moment(ev.startsAt).isBefore(eventStart) && moment(ev.endsAt).isAfter(eventEnd));
-      }).length;
-    }
-
-    function eventsComparer(a, b) {
-      var aStart = moment(a.startsAt);
-      var bStart = moment(b.startsAt);
-
-      if (aStart.isBefore(bStart)) {
-        return -1;
-      }
-
-      if (aStart.isSame(bStart)) {
-        var aEnd = moment(a.endsAt);
-        var bEnd = moment(b.endsAt);
-
-        if (aEnd.isSame(bEnd)) {
-          return 0;
-        } else if (aEnd.isAfter(bEnd)) {
-          return -1;
-        }
-        return 1;
-      }
-      return 1;
-    }
-
-    function getDayView(events, currentDay, dayViewStart, dayViewEnd, dayViewSplit, isWeekViewWithTimes) {
-      var baseBucketWidth = isWeekViewWithTimes ? 14.285714285714285 : 150;
       var dayStartHour = moment(dayViewStart || '00:00', 'HH:mm').hours();
       var dayEndHour = moment(dayViewEnd || '23:00', 'HH:mm').hours();
       var hourHeight = (60 / dayViewSplit) * 30;
-      var calendarStart = moment(currentDay).startOf('day').add(dayStartHour, 'hours');
-      var calendarEnd = moment(currentDay).startOf('day').add(dayEndHour, 'hours');
+      var calendarStart = moment(viewDate).startOf('day').add(dayStartHour, 'hours');
+      var calendarEnd = moment(viewDate).startOf('day').add(dayEndHour, 'hours');
       var calendarHeight = (dayEndHour - dayStartHour + 1) * hourHeight;
       var hourHeightMultiplier = hourHeight / 60;
       var buckets = [];
       var eventsInPeriod = filterEventsInPeriod(
         events,
-        moment(currentDay).startOf('day').toDate(),
-        moment(currentDay).endOf('day').toDate()
+        moment(viewDate).startOf('day').toDate(),
+        moment(viewDate).endOf('day').toDate()
       );
 
-      return eventsInPeriod.sort(eventsComparer).map(function(event) {
+      return eventsInPeriod.map(function(event) {
         if (moment(event.startsAt).isBefore(calendarStart)) {
           event.top = 0;
         } else {
@@ -300,10 +275,12 @@ angular
         }
 
         event.left = 0;
+
         return event;
       }).filter(function(event) {
         return event.height > 0;
       }).map(function(event) {
+
         var cannotFitInABucket = true;
         buckets.forEach(function(bucket, bucketIndex) {
           var canFitInThisBucket = true;
@@ -317,47 +294,36 @@ angular
 
           if (canFitInThisBucket && cannotFitInABucket) {
             cannotFitInABucket = false;
-            event.left = bucketIndex * baseBucketWidth;
-            if (isWeekViewWithTimes) {
-              event.bucketIndex = buckets.length;
-            }
+            event.left = bucketIndex * 150;
             buckets[bucketIndex].push(event);
           }
+
         });
 
         if (cannotFitInABucket) {
-          event.left = buckets.length * baseBucketWidth;
-          if (isWeekViewWithTimes) {
-            event.bucketIndex = buckets.length;
-          }
+          event.left = buckets.length * 150;
           buckets.push([event]);
         }
+
         return event;
-      }).map(function(event) {
-        if (isWeekViewWithTimes) {
-          event.width = getCrossingsCount(event, eventsInPeriod) > 0 ? baseBucketWidth / buckets.length : baseBucketWidth;
-          event.left = event.bucketIndex * baseBucketWidth / (buckets.length);
-          delete event.bucketIndex;
-        }
-        return event;
+
       });
+
     }
 
-    function getWeekViewWithTimes(events, currentDay, dayViewStart, dayViewEnd, dayViewSplit) {
-      var weekView = getWeekView(events, currentDay, false);
+    function getWeekViewWithTimes(events, viewDate, dayViewStart, dayViewEnd, dayViewSplit) {
+      var weekView = getWeekView(events, viewDate);
       var newEvents = [];
       weekView.days.forEach(function(day) {
         var dayEvents = weekView.events.filter(function(event) {
-          return moment(event.startsAt).isSame(moment(day.date), 'day') &&
-            moment(event.endsAt).isSame(moment(day.date), 'day');
+          return moment(event.startsAt).startOf('day').isSame(moment(day.date).startOf('day'));
         });
         var newDayEvents = getDayView(
           dayEvents,
           day.date,
           dayViewStart,
           dayViewEnd,
-          dayViewSplit,
-          true
+          dayViewSplit
         );
         newEvents = newEvents.concat(newDayEvents);
       });
@@ -372,6 +338,17 @@ angular
       return ((dayViewEndM.diff(dayViewStartM, 'hours') + 1) * hourHeight) + 2;
     }
 
+    function loadTemplates() {
+
+      var templatePromises = Object.keys(calendarConfig.templates).map(function(key) {
+        var templateUrl = calendarConfig.templates[key];
+        return $templateRequest(templateUrl);
+      });
+
+      return $q.all(templatePromises);
+
+    }
+
     return {
       getWeekDayNames: getWeekDayNames,
       getYearView: getYearView,
@@ -382,9 +359,8 @@ angular
       getDayViewHeight: getDayViewHeight,
       adjustEndDateFromStartDiff: adjustEndDateFromStartDiff,
       formatDate: formatDate,
-      eventIsInPeriod: eventIsInPeriod, //expose for testing only
-      getCrossingsCount: getCrossingsCount, //expose for testing only
-      eventsComparer: eventsComparer //expose for testing only
+      loadTemplates: loadTemplates,
+      eventIsInPeriod: eventIsInPeriod //expose for testing only
     };
 
   });