+++ test/unit/directives/mwlCalendarMonth.spec.js
@@ -11,11 +11,10 @@ describe('mwlCalendarMonth directive', function() {
     directiveScope,
     showModal,
     calendarHelper,
-    $templateCache,
     template =
       '<mwl-calendar-month ' +
       'events="events" ' +
-      'current-day="currentDay" ' +
+      'view-date="viewDate" ' +
       'on-event-click="onEventClick" ' +
       'on-event-times-changed="onEventTimesChanged" ' +
       'day-view-start="dayViewStart" ' +
@@ -30,7 +29,7 @@ describe('mwlCalendarMonth directive', function() {
 
   function prepareScope(vm) {
     //These variables MUST be set as a minimum for the calendar to work
-    vm.currentDay = calendarDay;
+    vm.viewDate = calendarDay;
     vm.cellIsOpen = true;
     vm.dayViewStart = '06:00';
     vm.dayViewEnd = '22:00';
@@ -81,9 +80,8 @@ describe('mwlCalendarMonth directive', function() {
 
   beforeEach(angular.mock.module('mwl.calendar'));
 
-  beforeEach(angular.mock.inject(function($compile, _$rootScope_, _calendarHelper_, _$templateCache_) {
+  beforeEach(angular.mock.inject(function($compile, _$rootScope_, _calendarHelper_) {
     $rootScope = _$rootScope_;
-    $templateCache = _$templateCache_;
     calendarHelper = _calendarHelper_;
     scope = $rootScope.$new();
     prepareScope(scope);
@@ -102,7 +100,7 @@ describe('mwlCalendarMonth directive', function() {
     sinon.stub(calendarHelper, 'getWeekViewWithTimes').returns({event: 'event2'});
     scope.$broadcast('calendar.refreshView');
     expect(calendarHelper.getWeekDayNames).to.have.been.called;
-    expect(calendarHelper.getMonthView).to.have.been.calledWith(scope.events, scope.currentDay);
+    expect(calendarHelper.getMonthView).to.have.been.calledWith(scope.events, scope.viewDate);
     expect(MwlCalendarCtrl.weekDays).to.eql(['Mon', 'Tu']);
     expect(MwlCalendarCtrl.view).to.equal(monthView);
     expect(MwlCalendarCtrl.openRowIndex).to.equal(0);
@@ -117,7 +115,8 @@ describe('mwlCalendarMonth directive', function() {
     expect(MwlCalendarCtrl.openDayIndex).to.equal(0);
     expect(showModal).to.have.been.calledWith('Day clicked', {
       calendarDate: MwlCalendarCtrl.view[0].date.toDate(),
-      $event: undefined
+      $event: undefined,
+      calendarCell: MwlCalendarCtrl.view[0]
     });
 
     //Close event list
@@ -127,11 +126,11 @@ describe('mwlCalendarMonth directive', function() {
   });
 
   it('should disable the slidebox if the click event is prevented', function() {
-    expect(MwlCalendarCtrl.openRowIndex).to.be.undefined;
+    expect(MwlCalendarCtrl.openRowIndex).to.be.null;
     expect(MwlCalendarCtrl.openDayIndex).to.be.undefined;
     MwlCalendarCtrl.view = [{date: moment(calendarDay), inMonth: true}];
     MwlCalendarCtrl.dayClicked(MwlCalendarCtrl.view[0], false, {defaultPrevented: true});
-    expect(MwlCalendarCtrl.openRowIndex).to.be.undefined;
+    expect(MwlCalendarCtrl.openRowIndex).to.be.null;
     expect(MwlCalendarCtrl.openDayIndex).to.be.undefined;
   });
 
@@ -156,44 +155,28 @@ describe('mwlCalendarMonth directive', function() {
   });
 
   it('should call the callback function when you finish dropping an event', function() {
-    MwlCalendarCtrl.handleEventDrop(scope.events[0], calendarDay);
+    var draggedFromDate = new Date();
+    MwlCalendarCtrl.handleEventDrop(scope.events[0], calendarDay, draggedFromDate);
     expect(showModal).to.have.been.calledWith('Dropped or resized', {
       calendarEvent: scope.events[0],
       calendarDate: new Date(2015, 4, 1),
       calendarNewEventStart: new Date(2015, 4, 1, 8, 0),
-      calendarNewEventEnd: new Date(2015, 4, 10, 9, 0)
+      calendarNewEventEnd: new Date(2015, 4, 10, 9, 0),
+      calendarDraggedFromDate: draggedFromDate
     });
   });
 
   it('should call the callback function when you finish dropping an event with no end date', function() {
     delete scope.events[0].endsAt;
-    MwlCalendarCtrl.handleEventDrop(scope.events[0], calendarDay);
+    var draggedFromDate = new Date();
+    MwlCalendarCtrl.handleEventDrop(scope.events[0], calendarDay, draggedFromDate);
     expect(showModal).to.have.been.calledWith('Dropped or resized', {
       calendarEvent: scope.events[0],
       calendarDate: new Date(2015, 4, 1),
       calendarNewEventStart: new Date(2015, 4, 1, 8, 0),
-      calendarNewEventEnd: null
+      calendarNewEventEnd: null,
+      calendarDraggedFromDate: draggedFromDate
     });
-  });
-
-  it('should use a custom cell url', function() {
-    var templatePath = 'customMonthCell.html';
-    $templateCache.put(templatePath, '<my-custom-cell>Hello world!</my-custom-cell>');
-    scope.monthCellTemplateUrl = templatePath;
-    MwlCalendarCtrl.cellModifier = angular.noop;
-    scope.$broadcast('calendar.refreshView');
-    scope.$apply();
-    expect(element.find('my-custom-cell').length).to.be.at.least(1);
-  });
-
-  it('should use a custom cell events url', function() {
-    var templatePath = 'customMonthCellEvents.html';
-    $templateCache.put(templatePath, '<my-custom-events>Hello world!</my-custom-events>');
-    scope.monthCellEventsTemplateUrl = templatePath;
-    MwlCalendarCtrl.cellModifier = angular.noop;
-    scope.$broadcast('calendar.refreshView');
-    scope.$apply();
-    expect(element.find('my-custom-events').length).to.be.at.least(1);
   });
 
 });