+++ test/unit/services/calendarHelper.spec.js
@@ -6,11 +6,13 @@ beforeEach(angular.mock.module('mwl.calendar'));
 
 describe('calendarHelper', function() {
 
-  var calendarHelper, events, clock, calendarDay, calendarConfig;
+  var calendarHelper, events, clock, calendarDay, calendarConfig, $templateCache, $rootScope;
 
-  beforeEach(inject(function(_calendarHelper_, _calendarConfig_) {
+  beforeEach(inject(function(_calendarHelper_, _calendarConfig_, _$templateCache_, _$rootScope_) {
     calendarHelper = _calendarHelper_;
     calendarConfig = _calendarConfig_;
+    $templateCache = _$templateCache_;
+    $rootScope = _$rootScope_;
 
     events = [{
       title: 'Event 1',
@@ -141,56 +143,6 @@ describe('calendarHelper', function() {
 
       expect(isInPeriod).to.be.true;
 
-    });
-
-  });
-
-  describe('getCrossingsCount', function() {
-    var dayEvents;
-
-    beforeEach(function() {
-      dayEvents = [{
-        $id: 1,
-        startsAt: new Date('November 11, 2015 02:00:00'),
-        endsAt: new Date('November 11, 2015 06:00:00')
-      }, {
-        $id: 2,
-        startsAt: new Date('November 11, 2015 06:00:00'),
-        endsAt: new Date('November 11, 2015 10:00:00')
-      }];
-    });
-
-    it('should have 1 crossing', function() {
-      var event = {
-        $id: 3,
-        startsAt: new Date('November 11, 2015 01:00:00'),
-        endsAt: new Date('November 11, 2015 03:00:00')
-      };
-      dayEvents.push(event);
-
-      expect(calendarHelper.getCrossingsCount(event, dayEvents)).to.eql(1);
-    });
-
-    it('should have 2 crossings', function() {
-      var event = {
-        $id: 3,
-        startsAt: new Date('November 11, 2015 05:00:00'),
-        endsAt: new Date('November 11, 2015 07:00:00')
-      };
-      dayEvents.push(event);
-
-      expect(calendarHelper.getCrossingsCount(event, dayEvents)).to.eql(2);
-    });
-
-    it('should have no crossings', function() {
-      var event = {
-        $id: 3,
-        startsAt: new Date('November 11, 2015 10:00:00'),
-        endsAt: new Date('November 11, 2015 12:00:00')
-      };
-      dayEvents.push(event);
-
-      expect(calendarHelper.getCrossingsCount(event, dayEvents)).to.eql(0);
     });
 
   });
@@ -528,6 +480,19 @@ describe('calendarHelper', function() {
 
     });
 
+    describe('recurring events', function() {
+
+      it('should display recuring events', function() {
+        weekView = calendarHelper.getWeekView([{
+          startsAt: new Date(2016, 0, 9, 1),
+          recursOn: 'month'
+        }], new Date(2016, 1, 9, 1));
+        expect(weekView.events[0].daySpan).to.equal(1);
+        expect(weekView.events[0].dayOffset).to.equal(2);
+      });
+
+    });
+
   });
 
   describe('getDayView', function() {
@@ -613,65 +578,20 @@ describe('calendarHelper', function() {
     });
   });
 
-  describe('eventsComparer', function() {
-    var a, b, c;
-
-    beforeEach(function() {
-      a = {
-        startsAt: new Date('October 20, 2015 11:00:00'),
-        endsAt: new Date('October 20, 2015 13:00:00')
-      };
-      b = {
-        startsAt: new Date('October 20, 2015 11:00:00'),
-        endsAt: new Date('October 20, 2015 12:00:00')
-      };
-      c = {
-        startsAt: new Date('October 20, 2015 11:00:00'),
-        endsAt: new Date('October 20, 2015 13:00:00')
-      };
-    });
-
-    it('eventA should be before eventB', function() {
-      expect(calendarHelper.eventsComparer(a, b)).to.equal(-1);
-    });
-
-    it('eventB should be after eventA', function() {
-      expect(calendarHelper.eventsComparer(b, a)).to.equal(1);
-    });
-
-    it('eventA should be equal eventC', function() {
-      expect(calendarHelper.eventsComparer(a, c)).to.equal(0);
-    });
-  });
-
   describe('getWeekViewWithTimes', function() {
-    var weekViewWithTimes, weekView;
+    var weekViewWithTimes;
 
     beforeEach(function() {
       var dayEvents = [{
-        $id: 1,
         startsAt: new Date('October 19, 2015 11:00:00'),
         endsAt: new Date('October 21, 2015 11:00:00')
       }, {
-        $id: 2,
         startsAt: new Date('October 20, 2015 11:00:00'),
         endsAt: new Date('October 21, 2015 11:00:00')
       }, {
-        $id: 3,
         startsAt: new Date('October 20, 2015 11:00:00'),
         endsAt: new Date('October 20, 2015 12:00:00')
-      },
-      {
-        $id: 4,
-        startsAt: new Date('October 20, 2015 11:00:00'),
-        endsAt: new Date('October 20, 2015 13:00:00')
       }];
-
-      weekView = calendarHelper.getWeekView(
-        dayEvents,
-        calendarDay,
-        true
-      );
 
       weekViewWithTimes = calendarHelper.getWeekViewWithTimes(
         dayEvents,
@@ -683,53 +603,38 @@ describe('calendarHelper', function() {
     });
 
     it('should calculate the week view with times', function() {
-      var baseBucketWidth = 14.285714285714285;
-
       var expectedEventsWeekView = [
         {
-          $id: 1,
           startsAt: new Date('October 19, 2015 11:00:00'),
           endsAt: new Date('October 21, 2015 11:00:00'),
           daySpan: 3,
-          dayOffset: 1
+          dayOffset: 1,
+          top: 658,
+          height: 782,
+          left: 0
         },
         {
-          $id: 2,
           startsAt: new Date('October 20, 2015 11:00:00'),
           endsAt: new Date('October 21, 2015 11:00:00'),
           daySpan: 2,
-          dayOffset: 2
-        }
-      ];
-
-      var expectedEventsWeekViewWithTimes = [
-        {
-          $id: 4,
-          startsAt: new Date('October 20, 2015 11:00:00'),
-          endsAt: new Date('October 20, 2015 13:00:00'),
-          daySpan: 1,
           dayOffset: 2,
           top: 658,
-          height: 120,
-          left: 0,
-          width: baseBucketWidth / 2
+          height: 782,
+          left: 0
         },
         {
-          $id: 3,
           startsAt: new Date('October 20, 2015 11:00:00'),
           endsAt: new Date('October 20, 2015 12:00:00'),
           daySpan: 1,
           dayOffset: 2,
           top: 658,
           height: 60,
-          left: baseBucketWidth / 2,
-          width: baseBucketWidth / 2
+          left: 150
         }
       ];
 
       expect(weekViewWithTimes.days.length).to.equal(7);
-      expect(weekView.events).to.eql(expectedEventsWeekView);
-      expect(weekViewWithTimes.events).to.eql(expectedEventsWeekViewWithTimes);
+      expect(weekViewWithTimes.events).to.eql(expectedEventsWeekView);
     });
   });
 
@@ -744,6 +649,22 @@ describe('calendarHelper', function() {
       calendarConfig.dateFormatter = 'moment';
       var formattedDate = calendarHelper.formatDate(new Date(), 'YYYY-MM-DD');
       expect(formattedDate).to.equal('2015-10-20');
+    });
+
+  });
+
+  describe('loadTemplates', function() {
+
+    it('should return a promise with all loaded templates', function(done) {
+      calendarConfig.templates = {
+        template: 'test.html'
+      };
+      $templateCache.put('test.html', 'contents');
+      calendarHelper.loadTemplates().then(function(result) {
+        expect(result).to.deep.equal(['contents']);
+        done();
+      });
+      $rootScope.$apply();
     });
 
   });